Move FIO address actions to a properly-named file
