+++ src/components/scenes/FioAddressListScene.js
@@ -6,11 +6,11 @@ import { ActivityIndicator, Image, ScrollView, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
+import { refreshAllFioAddresses } from '../../actions/FioAddressActions.js'
 import fioAddressLogo from '../../assets/images/fio/fio_logo.png'
 import { Fontello } from '../../assets/vector'
 import { FIO_ADDRESS_DETAILS, FIO_ADDRESS_REGISTER, FIO_DOMAIN_REGISTER, FIO_DOMAIN_SETTINGS } from '../../constants/SceneKeys.js'
 import s from '../../locales/strings.js'
-import { refreshAllFioAddresses } from '../../modules/FioAddress/action'
 import { FioNameRow } from '../../modules/FioAddress/components/FioName'
 import Gradient from '../../modules/UI/components/Gradient/Gradient.ui'
 import { PLATFORM } from '../../theme/variables/platform'
