+++ src/components/scenes/FioAddressRegisterScene.js
@@ -6,11 +6,11 @@ import { ActivityIndicator, Image, ScrollView, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import { sprintf } from 'sprintf-js'
 
+import { createFioWallet } from '../../actions/FioAddressActions.js'
 import { Fontello } from '../../assets/vector'
 import { FIO_ADDRESS_REGISTER_SELECT_WALLET, FIO_NAME_CONFIRM } from '../../constants/SceneKeys.js'
 import { CURRENCY_PLUGIN_NAMES, FIO_ADDRESS_DELIMITER, FIO_DOMAIN_DEFAULT, FIO_STR } from '../../constants/WalletAndCurrencyConstants.js'
 import s from '../../locales/strings.js'
-import { createFioWallet } from '../../modules/FioAddress/action'
 import { DomainListModal } from '../../modules/FioAddress/components/DomainListModal'
 import { checkIsDomainPublic } from '../../modules/FioAddress/util'
 import { connect } from '../../types/reactRedux.js'
