+++ src/components/scenes/FioAddressSettingsScene.js
@@ -4,10 +4,10 @@ import type { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 import { Actions } from 'react-native-router-flux'
 
-import { FIO_ADDRESS_LIST, SEND } from '../../constants/SceneKeys.js'
+import { refreshAllFioAddresses } from '../../actions/FioAddressActions.js'
+import { FIO_ADDRESS_LIST, SEND } from '../../constants/SceneKeys'
 import { formatDate } from '../../locales/intl.js'
 import s from '../../locales/strings'
-import { refreshAllFioAddresses } from '../../modules/FioAddress/action'
 import { FioActionSubmit } from '../../modules/FioAddress/components/FioActionSubmit'
 import { getRenewalFee, getTransferFee, renewFioName } from '../../modules/FioAddress/util'
 import { connect } from '../../types/reactRedux.js'
