+++ src/components/scenes/FioDomainRegisterScene.js
@@ -6,10 +6,10 @@ import { ActivityIndicator, ScrollView, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
+import { createFioWallet } from '../../actions/FioAddressActions.js'
 import { FIO_DOMAIN_REGISTER_SELECT_WALLET } from '../../constants/SceneKeys.js'
 import { CURRENCY_PLUGIN_NAMES, FIO_STR } from '../../constants/WalletAndCurrencyConstants.js'
 import s from '../../locales/strings.js'
-import { createFioWallet } from '../../modules/FioAddress/action.js'
 import { connect } from '../../types/reactRedux.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { SingleInputModal } from '../modals/SingleInputModal'
