+++ src/components/scenes/FioDomainSettingsScene.js
@@ -4,11 +4,11 @@ import type { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 import { Actions } from 'react-native-router-flux'
 
+import { refreshAllFioAddresses } from '../../actions/FioAddressActions.js'
 import { FIO_ADDRESS_LIST, SEND } from '../../constants/SceneKeys.js'
 import { FIO_ADDRESS_DELIMITER } from '../../constants/WalletAndCurrencyConstants.js'
 import { formatDate } from '../../locales/intl.js'
 import s from '../../locales/strings'
-import { refreshAllFioAddresses } from '../../modules/FioAddress/action'
 import { FioActionSubmit } from '../../modules/FioAddress/components/FioActionSubmit'
 import { getDomainSetVisibilityFee, getRenewalFee, getTransferFee, renewFioName, setDomainVisibility } from '../../modules/FioAddress/util'
 import { connect } from '../../types/reactRedux.js'
