+++ src/components/themed/SelectFioAddress.js
@@ -5,10 +5,10 @@ import * as React from 'react'
 import { View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 
+import { refreshAllFioAddresses } from '../../actions/FioAddressActions.js'
 import { FIO_ADDRESS_SETTINGS } from '../../constants/SceneKeys.js'
 import { FIO_STR } from '../../constants/WalletAndCurrencyConstants.js'
 import s from '../../locales/strings.js'
-import { refreshAllFioAddresses } from '../../modules/FioAddress/action.js'
 import { checkRecordSendFee, findWalletByFioAddress, FIO_NO_BUNDLED_ERR_CODE } from '../../modules/FioAddress/util.js'
 import { getSelectedWallet } from '../../selectors/WalletSelectors.js'
 import { connect } from '../../types/reactRedux.js'
